app/requirements.txt
click==8.1.3
Flask==2.2.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
peewee==3.16.0
PyJWT==2.6.0
Werkzeug==2.2.3
